// Jobs.Frontend/src/app/core/api/model/companyResponse.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';


/**
 * Response DTO for a company.
 */
export interface CompanyResponse { 
    id?: string;
    name?: string | null;
    location?: Location;
    website?: string | null;
}