// Jobs.Frontend/src/app/core/api/model/createCompanyRequest.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';


/**
 * Request DTO for creating a company.
 */
export interface CreateCompanyRequest { 
    name: string;
    location: Location;
    website?: string | null;
}