// Jobs.Frontend/src/app/core/api/model/updateCompanyRequest.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';


/**
 * Request DTO for updating a company.
 */
export interface UpdateCompanyRequest { 
    name: string;
    location: Location;
    website?: string | null;
}